+++ Report/2014_11235_LK.cpp
@@ -1,0 +1,51 @@
+#include "LK.h"
+
+// w3: left steer value
+// w4: right steer value
+
+void sim_main()
+{
+  double max_steering = 5000.0;
+
+  if (rtU->read2 <= 5000.0)
+  {
+    rtDW->w3 = 0.0;
+    rtDW->w4 = 0.0;
+  }
+  else
+  {
+    if (rtU->read1 > 0.0)
+    {
+      rtDW->w3 = 0.0;
+    }
+    else if ((rtU->read1 < 0.0) && (rtU->read1 < -max_steering))
+    {
+      rtDW->w3 = max_steering;
+    }
+    else
+    {
+      if ((rtU->read1 < 0.0) && (rtU->read1 >= -max_steering))
+      {
+        rtDW->w3 = rtU->read1 * -1;
+      }
+    }
+
+    if ((rtU->read1 > 0.0) && (rtU->read1 > max_steering))
+    {
+      rtDW->w4 = max_steering;
+    }
+    else if ((rtU->read1 > 0.0) && (rtU->read1 <= max_steering))
+    {
+      rtDW->w4 = rtU->read1;
+    }
+    else
+    {
+      if (rtU->read1 < 0.0)
+      {
+        rtDW->w4 = 0.0;
+      }
+    }
+  }
+  rtY->write3 = rtDW->w3;
+  rtY->write4 = rtDW->w4;
+}
