+++ Report/2014_11235_report1.md
@@ -1,0 +1,12 @@
+## Report 1
+
+### Modified Files
+CC.cpp: I have increased the max target speed to 100km/h
+LK.cpp: I have changed the max steering value to 5000
+design.xml: I have only modified LK and CC properties. I changed the period and deadline to 100, and execution time to 10.
+
+### Explanation
+When I changed the target speed to 100, the vehicle kept driving to the left or right and bumped into the wall. Therefore, I tried to change the maximum steering value to 5000, but it still did not changed the result. I noticed that the ECU is running in rate monotonic scheduling policy, so I wanted to prioritize CC and LK for better control of the vehicle. As a result, I changed the period of CC and LK from 500 to 100, making CC and LK first priority. Now, CC and LK can execute multiple times compared to the baseline configuration. 
+
+### Output
+The final best lap time is 56.6 seconds. I have reduced about 2 minutes from the baseline lap time.
