+++ design.xml
@@ -1,58 +1,48 @@
 <?xml version="1.0" encoding="UTF-8"?><hyundaiDesignTool>
     <ETHs>
-        <ETH IP="192.168.0.99">
+        <ETH IP="192.168.0.100">
             <link rid="ECU0" type="ECU"/>
-            <link rid="ECU1" type="ECU"/>
             <link rid="CAR" type="CAR"/>
         </ETH>
     </ETHs>
     <CAR IP="CAR">
-        <link rid="192.168.0.54" type="ETH"/>
+        <link rid="192.168.0.100" type="ETH"/>
     </CAR>
     <SWCs>
-        <SWC BCET="10" IP="SENSING" Virtual="1" WCET="10" consumer="" deadline="50" period="50" phase="0" readCon="0" writeCon="0">
-            <impl path="/home/park/git_base/CPSim_Linux_Generalized/sharedObjectFiles/SENSING/SENSING.cpp" type="MANUAL"/>
+        <SWC BCET="10" IP="LK" Virtual="1" WCET="10" consumer="" deadline="200" period="100" phase="0" readCon="1" writeCon="1">
+            <impl path="/home/ywj7373/CPSim_Linux_Generalized/sharedObjectFiles/LK/LK.cpp" type="MANUAL"/>
             <link idxCore="0" rid="ECU0" type="ECU"/>
         </SWC>
-        <SWC BCET="10" IP="LK" Virtual="1" WCET="10" consumer="" deadline="10" period="10" phase="0" readCon="1" writeCon="1">
-            <impl path="/home/park/git_base/CPSim_Linux_Generalized/sharedObjectFiles/LK/LK.cpp" type="MANUAL"/>
-            <link idxCore="0" rid="ECU1" type="ECU"/>
-        </SWC>
-        <SWC BCET="10" IP="CC" Virtual="1" WCET="10" consumer="" deadline="10" period="10" phase="0" readCon="1" writeCon="1">
-            <impl path="/home/park/git_base/CPSim_Linux_Generalized/sharedObjectFiles/CC/CC.cpp" type="MANUAL"/>
-            <link idxCore="0" rid="ECU2" type="ECU"/>
-        </SWC>
-        <SWC BCET="10" IP="BRAKE" Virtual="1" WCET="10" consumer="" deadline="50" period="50" phase="0" readCon="0" writeCon="0">
-            <impl path="/home/park/git_base/CPSim_Linux_Generalized/sharedObjectFiles/SENSING/SENSING.cpp" type="MANUAL"/>
+        <SWC BCET="10" IP="CC" Virtual="1" WCET="10" consumer="" deadline="200" period="100" phase="0" readCon="1" writeCon="1">
+            <impl path="/home/ywj7373/CPSim_Linux_Generalized/sharedObjectFiles/CC/CC.cpp" type="MANUAL"/>
             <link idxCore="0" rid="ECU0" type="ECU"/>
         </SWC>
-        <SWC BCET="10" IP="POWER" Virtual="1" WCET="10" consumer="" deadline="100" period="100" phase="0" readCon="0" writeCon="0">
-            <impl path="/home/park/Desktop/example.cpp" type="MANUAL"/>
+        <SWC BCET="100" IP="POWER" Virtual="1" WCET="100" consumer="" deadline="500" period="500" phase="0" readCon="0" writeCon="0">
+            <impl path="/home/ywj7373/CPSim_Linux_Generalized/sharedObjectFiles/POWER/POWER.cpp" type="MANUAL"/>
             <link idxCore="0" rid="ECU0" type="ECU"/>
         </SWC>
-        <SWC BCET="10" IP="LOCALIZATION" Virtual="1" WCET="10" consumer="" deadline="100" period="100" phase="0" readCon="0" writeCon="0">
-            <impl path="/home/park/Desktop/example.cpp" type="MANUAL"/>
+        <SWC BCET="100" IP="BRAKE" Virtual="1" WCET="100" consumer="" deadline="500" period="500" phase="0" readCon="0" writeCon="0">
+            <impl path="/home/ywj7373/CPSim_Linux_Generalized/sharedObjectFiles/BRAKE/BRAKE.cpp" type="MANUAL"/>
+            <link idxCore="0" rid="ECU0" type="ECU"/>
+        </SWC>
+        <SWC BCET="100" IP="LOCAL" Virtual="1" WCET="100" consumer="" deadline="500" period="500" phase="0" readCon="0" writeCon="0">
+            <impl path="/home/ywj7373/CPSim_Linux_Generalized/sharedObjectFiles/LOCAL/LOCAL.cpp" type="MANUAL"/>
+            <link idxCore="0" rid="ECU0" type="ECU"/>
+        </SWC>
+        <SWC BCET="100" IP="SENSING" Virtual="1" WCET="100" consumer="" deadline="500" period="500" phase="0" readCon="0" writeCon="0">
+            <impl path="/home/ywj7373/CPSim_Linux_Generalized/sharedObjectFiles/SENSING/SENSING.cpp" type="MANUAL"/>
             <link idxCore="0" rid="ECU0" type="ECU"/>
         </SWC>
     </SWCs>
     <ECUs>
         <ECU IP="ECU0" Virtual="1" numCores="1" schedPolicy="RM" sysClock="100" targetArch="SPC56EL70">
             <link idxCore="0" rid="LK" type="SWC"/>
+            <link idxCore="0" rid="CC" type="SWC"/>
+            <link idxCore="0" rid="POWER" type="SWC"/>
+            <link idxCore="0" rid="BRAKE" type="SWC"/>
             <link idxCore="0" rid="SENSING" type="SWC"/>
-            <link idxCore="0" rid="LOCALIZATION" type="SWC"/>
-            <link rid="192.168.0.54" type="ETH"/>
-        </ECU>
-        <ECU IP="ECU1" Virtual="1" numCores="1" schedPolicy="RM" sysClock="100" targetArch="SPC56EL70">
-            <link idxCore="0" rid="CC" type="SWC"/>
-            <link idxCore="0" rid="BRAKE" type="SWC"/>
-            <link idxCore="0" rid="POWER" type="SWC"/>
-            <link rid="192.168.0.54" type="ETH"/>
-        </ECU>
-        <ECU IP="ECU2" Virtual="1" numCores="1" schedPolicy="RM" sysClock="100" targetArch="SPC56EL70">
-            <link idxCore="0" rid="CC" type="SWC"/>
-            <link idxCore="0" rid="BRAKE" type="SWC"/>
-            <link idxCore="0" rid="POWER" type="SWC"/>
-            <link rid="192.168.0.54" type="ETH"/>
+            <link idxCore="0" rid="LOCAL" type="SWC"/>
+            <link rid="192.168.0.100" type="ETH"/>
         </ECU>
     </ECUs>
 </hyundaiDesignTool>
