+++ sharedObjectFiles/CC.cpp
@@ -62,7 +62,7 @@ void sim_main()
   /********************************************/
   write5 = 0;
   write6 = 0;
-  read2 = read2 / 100 + 600000;
+  read2 = read2 / 100 + 200000;
 
   if(read4 > 6000) {
