+++ sharedObjectFiles/CC/CC.cpp
@@ -22,7 +22,7 @@ void sim_main()
   /********************************************/
   write5 = 0;
   write6 = 0;
-  read2 = read2 / 100 + 600000;
+  read2 = read2 / 100 + 200000;
 
   if(read4 > 6000) {
