+++ sharedObjectFiles/LK/LK.cpp
@@ -1,7 +1,12 @@
 #include "LK.h"
 
+// w3: left steer value
+// w4: right steer value
+
 void sim_main()
 {
+  double max_steering = 5000.0;
+
   if (rtU->read2 <= 5000.0)
   {
     rtDW->w3 = 0.0;
@@ -13,23 +18,23 @@ void sim_main()
     {
       rtDW->w3 = 0.0;
     }
-    else if ((rtU->read1 < 0.0) && (rtU->read1 < -10000.0))
+    else if ((rtU->read1 < 0.0) && (rtU->read1 < -max_steering))
     {
-      rtDW->w3 = 10000.0;
+      rtDW->w3 = max_steering;
     }
     else
     {
-      if ((rtU->read1 < 0.0) && (rtU->read1 >= -10000.0))
+      if ((rtU->read1 < 0.0) && (rtU->read1 >= -max_steering))
       {
         rtDW->w3 = rtU->read1 * -1;
       }
     }
 
-    if ((rtU->read1 > 0.0) && (rtU->read1 > 10000.0))
+    if ((rtU->read1 > 0.0) && (rtU->read1 > max_steering))
     {
-      rtDW->w4 = 10000.0;
+      rtDW->w4 = max_steering;
     }
-    else if ((rtU->read1 > 0.0) && (rtU->read1 <= 10000.0))
+    else if ((rtU->read1 > 0.0) && (rtU->read1 <= max_steering))
     {
       rtDW->w4 = rtU->read1;
     }
